Add file system wrapper and remove hook for snapshot.
Snapshot hook is hard to maintain and don't support to hook write progress. Acturally, we need file system like interfaces, to customize read and write operations, if needed.
Furthermore, a new option filter_before_copy_remote is provided. If this option is enable, checksum are used to filter duplicate files between local and remote, and the duplicate files will not be transmitted.

Change-Id: I54669d9979a710e8c97a05798f5ea76c4fd143cd