Fix a date consitency problem that caused by the wrong read size, returned by get_file(), in the process of installing snapshot.
Consider a system, in which the snapshot files may contain holes. In the process of installing snapshot, the follower acquire to read 1MB data from the leader, but the leader only get 512KB quota after throttle. If the 512KB just be a hole, the read size returned to the follower is zero! In this case, the follower think all the 1MB data are read, and the remain 512KB is skipped. The data consistency is broken, if there are valid data in the remain 512KB!