Set default value of raft_step_down_when_vote_timedout to true to avoid election failure in extreme case
A known case: in a raft group {a, b, c}, there is network problem
between a and b. For some reason a and b reach election timedout
successively, and get successful prevote response from c to become
Candidates successively. Then a get successful vote response from c
and become leader which writes a log in a and c. After that point,
b will fail to get successful vote from c because of its log is fewer
than c. Then b reaches vote timedout and raise its term to vote again,
but always fails because of log lack with the affect of raising c's
term as well, which will make leader a step_down to follower. This bad
case may contunue if b's vote timer is always before c's election timer.
To make b step_down to follower immediately when vote timedout FIX this
case by making a or c become leader of the group.