Self-protective readonly mode.
If a follower become readonly, the leader stop replicate new logs to it. This
may cause the data for behind the leader, in the case that the leader is still
writable. After the follower exit readonly mode, the leader will resume to
replicate missing logs.

A leader is readonly, if the node itself is readonly, or writable nodes (nodes that
are not marked as readonly) in the group is less than majority. Once a leader become
readonly, no new users logs will be acceptted.